:recycle: REFACTOR: AuthService 인터페이스 제거 (#11)
